Secound Update

+++ app/Http/Controllers/BooksController.php
@@ -10,6 +10,9 @@ use File;
 
 class BooksController extends Controller
 {
+    public function __construct() {
+        $this->middleware('auth', ['except' => ['index']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -48,12 +51,14 @@ class BooksController extends Controller
         if($request->hasFile('image')){
             $filename = str_random(10) . '.' . $request->file('image')->getClientOriginalExtension();
             $request->file('image')->move(public_path() . '/images/', $filename);
-            Image::make(public_path() . '/images/' . $filename)->resize(50, 50)->save(public_path() . '/images/resize/' . $filename);
+            Image::make(public_path() . '/images/' . $filename)->resize(100, 100)->save(public_path() . '/images/resize/' . $filename);
             $book->image = $filename;
         } else {
             $book->image = 'nopic.jpg';
         }
         $book->save();
+
+        $request->session()->flash('status', 'บันทึกข้อมูลเรียบร้อยแล้ว');
         return redirect()->action('BooksController@index');
         
     }
@@ -97,12 +102,20 @@ class BooksController extends Controller
         $book->price = $request->price;
         $book->typebooks_id = $request->typebooks_id;
 
-        if($book->image != 'nopic.jpg') {
-            File::delete(public_path().'\\images\\'.$book->image);
-            File::delete(public_path().'\\images\\resize\\'.$book->image);
+        if($request->hasFile('image')) {
+
+            if($book->image != 'nopic.jpg') {
+                File::delete(public_path().'\\images\\'.$book->image);
+                File::delete(public_path().'\\images\\resize\\'.$book->image);
+            }
+            $filename = str_random(10).'.'.$request->file('image')->getClientOriginalExtension();
+            $request->file('image')->move(public_path().'/images/', $filename);
+            Image::make(public_path(). '/images/'.$filename)->resize(100, 100)->save(public_path().'/images/resize/'.$filename);
+            $book->image = $filename;
         }
 
-        $file = str_random
+        $book->save();
+        return redirect()->action('BooksController@index');
     }
 
     /**
@@ -113,6 +126,13 @@ class BooksController extends Controller
      */
     public function destroy($id)
     {
-        //
+        $book = Books::find($id);
+        if($book->image != 'nopic.jpg') {
+            File::delete(public_path(). '\\images\\'.$book->image);
+            File::delete(public_path(). '\\images\\resize\\'.$book->image);
+        }
+
+        $book->delete();
+        return redirect()->action('BooksController@index');
     }
 }
